pyproject.toml
[tool.poetry]
name = "hacktor-app"
version = "0.1.0"
description = ""
authors = ["jchauhan <jitendra.chauhan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.40.2"
python-dotenv = "^1.0.1"
gradio = "^5.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
